Refactor out a Workout Class (#1)
* added WokrkoutFragment tests

* deleted example tests

* update github actions to run on dev branch

* update github actions to run on dev branch

* add coroutine dependencies

* add mockito dependency

* Create workout class for each session.

TODO integrate this into the DataModel.

Notes on dev choices are in (./notes.md).

* switch workout class to livedata

* Encapsulate stopwatch inside workout class.

Refactors the Stopwatch instance and the related distance fields out
of the ViewModel and into its own object. This makes it easier to add
in the speed calculations later.

Workout will be the one main class for each session. Later, these
should be allowed to handle the following:
- Caching on device while in progress (Room DB)
- Save to cloud (Firebase?)